makefile

+++ Makefile
@@ -3,10 +3,8 @@ APP_IMAGE_NAME := golang-tool-sandbox
 APP_CONTAINER_NAME := sandox
 
 docker-clean:
-#ifneq ($(docker ps -a -q),)
 	docker stop $$(docker ps -a -q)
-	docker rm  $$(docker ps -a -q) 
-#endif
+	docker rm  $$(docker ps -a -q)
 
 docker-build:
 	docker build -t $(APP_IMAGE_NAME) .
@@ -18,23 +16,20 @@ docker-run: docker-build
 BUILD_DIR := ./build
 
 clean:
-	rm -R $(BUILD_DIR)
+	if [ -d $(BUILD_DIR) ]; then rm -R $(BUILD_DIR); fi
 	mkdir $(BUILD_DIR)
 list-imports:
 	go list -f '{{ join .Imports "\n" }}'
-doc:
-	#go doc main
 wrk:
-	go-wrk -d 20 http://localhost:8080/hello/scristofari
+	# go run main.go
+	go-wrk -d 10 http://localhost:8080/hello/scristofari
+	# kill $!
 cover:
 	go test -coverprofile $(BUILD_DIR)/cover.out
 	go tool cover -html=$(BUILD_DIR)/cover.out -o $(BUILD_DIR)/cover.html
 pprof:
-	# run / wrk / firefox / graphviz
-	# command top / web
 	go tool pprof --seconds=5 http://localhost:8080/debug/pprof/profile
 torch:
-	# run / wrk
 	go-torch -t 5 --print > $(BUILD_DIR)/torch.svg
 bench:
 	go test -bench .
@@ -42,4 +37,9 @@ bench-profile:
 	go test -bench . -cpuprofile=$(BUILD_DIR)/prof.cpu -o $(BUILD_DIR)/golang-tool-sandbox.test
 torch-profile: bench-profile
 	go-torch --binaryname $(BUILD_DIR)/golang-tool-sandbox.test -b $(BUILD_DIR)/prof.cpu --print > $(BUILD_DIR)/torch-profile.svg
-all: clean cover torch-profile
+lint:
+	golint -set_exit_status .
+metalinter:
+	gometalinter --deadline=1m --cyclo-over=20 --tests --json . > $(BUILD_DIR)/linter.json
+
+all: clean lint metalinter cover torch-profile
